Split the last three screens into their own packages
Finishes the ui/word/ layout started in #860. ListDetailScreen (1125),
LanguageSetupScreen (962) and DownloadScreen (774) were the monoliths
left in ui/; they are smaller and more uniform than the first five, so
they go in one change rather than three.

  ui/listdetail/ListDetailViewModel.kt        522
  ui/listdetail/ListDetailScreen.kt           538
  ui/listdetail/preview/ListDetailPreviews.kt 303

  ui/languagesetup/LanguageSetupViewModel.kt        303
  ui/languagesetup/LanguageSetupScreen.kt           284
  ui/languagesetup/LanguageSetupComponents.kt       483
  ui/languagesetup/preview/LanguageSetupPreviews.kt 142

  ui/download/DownloadViewModel.kt        314
  ui/download/DownloadScreen.kt           439
  ui/download/preview/DownloadPreviews.kt 167

DownloadUiState moves with the ViewModel rather than the screen: it is
the state the ViewModel drives, and the screen only renders it.

Pure move, verified per screen by diffing the moved bodies against the
originals:

  Download        identical, no visibility changes at all
  ListDetail      one widening  (ListDetailErrorScreen, used by previews)
  LanguageSetup   six widenings (SectionVisualState and the row/section
                  composables the screen shares with its components file)

As in the Stats split, the widening list came from the compiler rather
than being guessed up front.

Two callers follow their symbols: WordListIcon imports WordListEmblem
from listdetail, and ListDetailViewModelTest imports the package it
exercises.

ui/ is now 18 files with no monolith left; the largest is
LanguageRequestDialog at 381 lines. Across the six splits it went from 29
files and eight screens over 750 lines to 18 files and none.
