+++ composeApp/src/commonMain/kotlin/com/slovy/slovymovyapp/App.kt
@@ -42,6 +42,9 @@ import com.slovy.slovymovyapp.ui.settings.*
 import com.slovy.slovymovyapp.ui.search.*
 import com.slovy.slovymovyapp.ui.developer.*
 import com.slovy.slovymovyapp.ui.stats.*
+import com.slovy.slovymovyapp.ui.listdetail.*
+import com.slovy.slovymovyapp.ui.languagesetup.*
+import com.slovy.slovymovyapp.ui.download.*
 import com.slovy.slovymovyapp.data.lists.ListsService
 import com.slovy.slovymovyapp.data.lists.WordListsRepository
 import com.slovy.slovymovyapp.data.recovery.RecoverableSense
